Driver for Cray ClusterStor E1000 NVMe slot LEDs
This patch introduces a driver to control the NVMe slot LEDs on the Cray
ClusterStore E1000.  It provides hotplug attention status callbacks for
the 24 NVMe slots on the E1000.  This allows users to access the E1000's
locate and fault LEDs via the normal /sys/bus/pci/slots/<slot>/attention
sysfs entries.  This driver uses IPMI to communicate with the E1000
controller to toggle the LEDs.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>